gh-142 Repackage jooq generated code
