+++ stroom-persistence/src/main/java/stroom/auth/db/Auth.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth;
+package stroom.auth.db;
 
 
 import java.util.ArrayList;
@@ -14,11 +14,10 @@ import org.jooq.Catalog;
 import org.jooq.Table;
 import org.jooq.impl.SchemaImpl;
 
-import stroom.db.DefaultCatalog;
-import stroom.db.auth.tables.JsonWebKey;
-import stroom.db.auth.tables.TokenTypes;
-import stroom.db.auth.tables.Tokens;
-import stroom.db.auth.tables.Users;
+import stroom.auth.db.tables.JsonWebKey;
+import stroom.auth.db.tables.TokenTypes;
+import stroom.auth.db.tables.Tokens;
+import stroom.auth.db.tables.Users;
 
 
 /**
@@ -34,7 +33,7 @@ import stroom.db.auth.tables.Users;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class Auth extends SchemaImpl {
 
-    private static final long serialVersionUID = -2079851059;
+    private static final long serialVersionUID = -1009352799;
 
     /**
      * The reference instance of <code>auth</code>
@@ -44,22 +43,22 @@ public class Auth extends SchemaImpl {
     /**
      * The table <code>auth.json_web_key</code>.
      */
-    public final JsonWebKey JSON_WEB_KEY = stroom.db.auth.tables.JsonWebKey.JSON_WEB_KEY;
+    public final JsonWebKey JSON_WEB_KEY = stroom.auth.db.tables.JsonWebKey.JSON_WEB_KEY;
 
     /**
      * The table <code>auth.tokens</code>.
      */
-    public final Tokens TOKENS = stroom.db.auth.tables.Tokens.TOKENS;
+    public final Tokens TOKENS = stroom.auth.db.tables.Tokens.TOKENS;
 
     /**
      * The table <code>auth.token_types</code>.
      */
-    public final TokenTypes TOKEN_TYPES = stroom.db.auth.tables.TokenTypes.TOKEN_TYPES;
+    public final TokenTypes TOKEN_TYPES = stroom.auth.db.tables.TokenTypes.TOKEN_TYPES;
 
     /**
      * The table <code>auth.users</code>.
      */
-    public final Users USERS = stroom.db.auth.tables.Users.USERS;
+    public final Users USERS = stroom.auth.db.tables.Users.USERS;
 
     /**
      * No further instances allowed
