+++ stroom-persistence/src/main/java/stroom/auth/db/DefaultCatalog.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db;
+package stroom.auth.db;
 
 
 import java.util.ArrayList;
@@ -12,8 +12,6 @@ import javax.annotation.Generated;
 
 import org.jooq.Schema;
 import org.jooq.impl.CatalogImpl;
-
-import stroom.db.auth.Auth;
 
 
 /**
@@ -29,7 +27,7 @@ import stroom.db.auth.Auth;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class DefaultCatalog extends CatalogImpl {
 
-    private static final long serialVersionUID = 2096972298;
+    private static final long serialVersionUID = -253301902;
 
     /**
      * The reference instance of <code></code>
@@ -39,7 +37,7 @@ public class DefaultCatalog extends CatalogImpl {
     /**
      * The schema <code>auth</code>.
      */
-    public final Auth AUTH = stroom.db.auth.Auth.AUTH;
+    public final Auth AUTH = stroom.auth.db.Auth.AUTH;
 
     /**
      * No further instances allowed
