+++ stroom-persistence/src/main/java/stroom/auth/db/Indexes.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth;
+package stroom.auth.db;
 
 
 import javax.annotation.Generated;
@@ -10,10 +10,10 @@ import org.jooq.Index;
 import org.jooq.OrderField;
 import org.jooq.impl.Internal;
 
-import stroom.db.auth.tables.JsonWebKey;
-import stroom.db.auth.tables.TokenTypes;
-import stroom.db.auth.tables.Tokens;
-import stroom.db.auth.tables.Users;
+import stroom.auth.db.tables.JsonWebKey;
+import stroom.auth.db.tables.TokenTypes;
+import stroom.auth.db.tables.Tokens;
+import stroom.auth.db.tables.Users;
 
 
 /**
