+++ stroom-persistence/src/main/java/stroom/auth/db/Keys.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth;
+package stroom.auth.db;
 
 
 import javax.annotation.Generated;
@@ -11,14 +11,14 @@ import org.jooq.Identity;
 import org.jooq.UniqueKey;
 import org.jooq.impl.Internal;
 
-import stroom.db.auth.tables.JsonWebKey;
-import stroom.db.auth.tables.TokenTypes;
-import stroom.db.auth.tables.Tokens;
-import stroom.db.auth.tables.Users;
-import stroom.db.auth.tables.records.JsonWebKeyRecord;
-import stroom.db.auth.tables.records.TokenTypesRecord;
-import stroom.db.auth.tables.records.TokensRecord;
-import stroom.db.auth.tables.records.UsersRecord;
+import stroom.auth.db.tables.JsonWebKey;
+import stroom.auth.db.tables.TokenTypes;
+import stroom.auth.db.tables.Tokens;
+import stroom.auth.db.tables.Users;
+import stroom.auth.db.tables.records.JsonWebKeyRecord;
+import stroom.auth.db.tables.records.TokenTypesRecord;
+import stroom.auth.db.tables.records.TokensRecord;
+import stroom.auth.db.tables.records.UsersRecord;
 
 
 /**
@@ -89,9 +89,9 @@ public class Keys {
     }
 
     private static class ForeignKeys0 {
-        public static final ForeignKey<TokensRecord, UsersRecord> FK_ISSUED_TO = Internal.createForeignKey(stroom.db.auth.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_issued_to", Tokens.TOKENS.USER_ID);
-        public static final ForeignKey<TokensRecord, TokenTypesRecord> FK_TOKEN_TYPE_ID = Internal.createForeignKey(stroom.db.auth.Keys.KEY_TOKEN_TYPES_PRIMARY, Tokens.TOKENS, "fk_token_type_id", Tokens.TOKENS.TOKEN_TYPE_ID);
-        public static final ForeignKey<TokensRecord, UsersRecord> FK_ISSUED_BY_USER = Internal.createForeignKey(stroom.db.auth.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_issued_by_user", Tokens.TOKENS.ISSUED_BY_USER);
-        public static final ForeignKey<TokensRecord, UsersRecord> FK_UPDATED_BY_USER = Internal.createForeignKey(stroom.db.auth.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_updated_by_user", Tokens.TOKENS.UPDATED_BY_USER);
+        public static final ForeignKey<TokensRecord, UsersRecord> FK_ISSUED_TO = Internal.createForeignKey(stroom.auth.db.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_issued_to", Tokens.TOKENS.USER_ID);
+        public static final ForeignKey<TokensRecord, TokenTypesRecord> FK_TOKEN_TYPE_ID = Internal.createForeignKey(stroom.auth.db.Keys.KEY_TOKEN_TYPES_PRIMARY, Tokens.TOKENS, "fk_token_type_id", Tokens.TOKENS.TOKEN_TYPE_ID);
+        public static final ForeignKey<TokensRecord, UsersRecord> FK_ISSUED_BY_USER = Internal.createForeignKey(stroom.auth.db.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_issued_by_user", Tokens.TOKENS.ISSUED_BY_USER);
+        public static final ForeignKey<TokensRecord, UsersRecord> FK_UPDATED_BY_USER = Internal.createForeignKey(stroom.auth.db.Keys.KEY_USERS_PRIMARY, Tokens.TOKENS, "fk_updated_by_user", Tokens.TOKENS.UPDATED_BY_USER);
     }
 }
