+++ stroom-persistence/src/main/java/stroom/auth/db/Tables.java
@@ -1,15 +1,15 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth;
+package stroom.auth.db;
 
 
 import javax.annotation.Generated;
 
-import stroom.db.auth.tables.JsonWebKey;
-import stroom.db.auth.tables.TokenTypes;
-import stroom.db.auth.tables.Tokens;
-import stroom.db.auth.tables.Users;
+import stroom.auth.db.tables.JsonWebKey;
+import stroom.auth.db.tables.TokenTypes;
+import stroom.auth.db.tables.Tokens;
+import stroom.auth.db.tables.Users;
 
 
 /**
@@ -28,20 +28,20 @@ public class Tables {
     /**
      * The table <code>auth.json_web_key</code>.
      */
-    public static final JsonWebKey JSON_WEB_KEY = stroom.db.auth.tables.JsonWebKey.JSON_WEB_KEY;
+    public static final JsonWebKey JSON_WEB_KEY = stroom.auth.db.tables.JsonWebKey.JSON_WEB_KEY;
 
     /**
      * The table <code>auth.tokens</code>.
      */
-    public static final Tokens TOKENS = stroom.db.auth.tables.Tokens.TOKENS;
+    public static final Tokens TOKENS = stroom.auth.db.tables.Tokens.TOKENS;
 
     /**
      * The table <code>auth.token_types</code>.
      */
-    public static final TokenTypes TOKEN_TYPES = stroom.db.auth.tables.TokenTypes.TOKEN_TYPES;
+    public static final TokenTypes TOKEN_TYPES = stroom.auth.db.tables.TokenTypes.TOKEN_TYPES;
 
     /**
      * The table <code>auth.users</code>.
      */
-    public static final Users USERS = stroom.db.auth.tables.Users.USERS;
+    public static final Users USERS = stroom.auth.db.tables.Users.USERS;
 }
