+++ stroom-persistence/src/main/java/stroom/auth/db/tables/JsonWebKey.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables;
+package stroom.auth.db.tables;
 
 
 import java.util.Arrays;
@@ -22,10 +22,10 @@ import org.jooq.UniqueKey;
 import org.jooq.impl.DSL;
 import org.jooq.impl.TableImpl;
 
-import stroom.db.auth.Auth;
-import stroom.db.auth.Indexes;
-import stroom.db.auth.Keys;
-import stroom.db.auth.tables.records.JsonWebKeyRecord;
+import stroom.auth.db.Auth;
+import stroom.auth.db.Indexes;
+import stroom.auth.db.Keys;
+import stroom.auth.db.tables.records.JsonWebKeyRecord;
 
 
 /**
@@ -41,7 +41,7 @@ import stroom.db.auth.tables.records.JsonWebKeyRecord;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class JsonWebKey extends TableImpl<JsonWebKeyRecord> {
 
-    private static final long serialVersionUID = 98995821;
+    private static final long serialVersionUID = -305415719;
 
     /**
      * The reference instance of <code>auth.json_web_key</code>
