+++ stroom-persistence/src/main/java/stroom/auth/db/tables/TokenTypes.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables;
+package stroom.auth.db.tables;
 
 
 import java.util.Arrays;
@@ -22,10 +22,10 @@ import org.jooq.UniqueKey;
 import org.jooq.impl.DSL;
 import org.jooq.impl.TableImpl;
 
-import stroom.db.auth.Auth;
-import stroom.db.auth.Indexes;
-import stroom.db.auth.Keys;
-import stroom.db.auth.tables.records.TokenTypesRecord;
+import stroom.auth.db.Auth;
+import stroom.auth.db.Indexes;
+import stroom.auth.db.Keys;
+import stroom.auth.db.tables.records.TokenTypesRecord;
 
 
 /**
@@ -41,7 +41,7 @@ import stroom.db.auth.tables.records.TokenTypesRecord;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class TokenTypes extends TableImpl<TokenTypesRecord> {
 
-    private static final long serialVersionUID = 2004925719;
+    private static final long serialVersionUID = -1916853629;
 
     /**
      * The reference instance of <code>auth.token_types</code>
