+++ stroom-persistence/src/main/java/stroom/auth/db/tables/Tokens.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables;
+package stroom.auth.db.tables;
 
 
 import java.sql.Timestamp;
@@ -23,10 +23,10 @@ import org.jooq.UniqueKey;
 import org.jooq.impl.DSL;
 import org.jooq.impl.TableImpl;
 
-import stroom.db.auth.Auth;
-import stroom.db.auth.Indexes;
-import stroom.db.auth.Keys;
-import stroom.db.auth.tables.records.TokensRecord;
+import stroom.auth.db.Auth;
+import stroom.auth.db.Indexes;
+import stroom.auth.db.Keys;
+import stroom.auth.db.tables.records.TokensRecord;
 
 
 /**
@@ -42,7 +42,7 @@ import stroom.db.auth.tables.records.TokensRecord;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class Tokens extends TableImpl<TokensRecord> {
 
-    private static final long serialVersionUID = 2003358541;
+    private static final long serialVersionUID = 715091897;
 
     /**
      * The reference instance of <code>auth.tokens</code>
