+++ stroom-persistence/src/main/java/stroom/auth/db/tables/Users.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables;
+package stroom.auth.db.tables;
 
 
 import java.sql.Timestamp;
@@ -23,10 +23,10 @@ import org.jooq.UniqueKey;
 import org.jooq.impl.DSL;
 import org.jooq.impl.TableImpl;
 
-import stroom.db.auth.Auth;
-import stroom.db.auth.Indexes;
-import stroom.db.auth.Keys;
-import stroom.db.auth.tables.records.UsersRecord;
+import stroom.auth.db.Auth;
+import stroom.auth.db.Indexes;
+import stroom.auth.db.Keys;
+import stroom.auth.db.tables.records.UsersRecord;
 
 
 /**
@@ -42,7 +42,7 @@ import stroom.db.auth.tables.records.UsersRecord;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class Users extends TableImpl<UsersRecord> {
 
-    private static final long serialVersionUID = 1756787340;
+    private static final long serialVersionUID = 1543983198;
 
     /**
      * The reference instance of <code>auth.users</code>
@@ -135,7 +135,7 @@ public class Users extends TableImpl<UsersRecord> {
     /**
      * The column <code>auth.users.never_expires</code>.
      */
-    public final TableField<UsersRecord, Boolean> NEVER_EXPIRES = createField("never_expires", org.jooq.impl.SQLDataType.BIT.defaultValue(org.jooq.impl.DSL.inline("b'0'", org.jooq.impl.SQLDataType.BIT)), this, "");
+    public final TableField<UsersRecord, Boolean> NEVER_EXPIRES = createField("never_expires", org.jooq.impl.SQLDataType.BOOLEAN.nullable(false).defaultValue(org.jooq.impl.DSL.inline("0", org.jooq.impl.SQLDataType.BOOLEAN)), this, "");
 
     /**
      * The column <code>auth.users.reactivated_date</code>.
