+++ stroom-persistence/src/main/java/stroom/auth/db/tables/records/JsonWebKeyRecord.java
@@ -1,18 +1,18 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables.records;
+package stroom.auth.db.tables.records;
 
 
 import javax.annotation.Generated;
 
 import org.jooq.Field;
 import org.jooq.Record1;
-import org.jooq.Record2;
-import org.jooq.Row2;
+import org.jooq.Record3;
+import org.jooq.Row3;
 import org.jooq.impl.UpdatableRecordImpl;
 
-import stroom.db.auth.tables.TokenTypes;
+import stroom.auth.db.tables.JsonWebKey;
 
 
 /**
@@ -26,36 +26,50 @@ import stroom.db.auth.tables.TokenTypes;
     comments = "This class is generated by jOOQ"
 )
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> implements Record2<Integer, String> {
+public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> implements Record3<Integer, String, String> {
 
-    private static final long serialVersionUID = -194124216;
+    private static final long serialVersionUID = 1425006793;
 
     /**
-     * Setter for <code>auth.token_types.id</code>.
+     * Setter for <code>auth.json_web_key.id</code>.
      */
     public void setId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for <code>auth.token_types.id</code>.
+     * Getter for <code>auth.json_web_key.id</code>.
      */
     public Integer getId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for <code>auth.token_types.token_type</code>.
+     * Setter for <code>auth.json_web_key.keyId</code>.
      */
-    public void setTokenType(String value) {
+    public void setKeyid(String value) {
         set(1, value);
     }
 
     /**
-     * Getter for <code>auth.token_types.token_type</code>.
+     * Getter for <code>auth.json_web_key.keyId</code>.
      */
-    public String getTokenType() {
+    public String getKeyid() {
         return (String) get(1);
+    }
+
+    /**
+     * Setter for <code>auth.json_web_key.json</code>.
+     */
+    public void setJson(String value) {
+        set(2, value);
+    }
+
+    /**
+     * Getter for <code>auth.json_web_key.json</code>.
+     */
+    public String getJson() {
+        return (String) get(2);
     }
 
     // -------------------------------------------------------------------------
@@ -71,23 +85,23 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
     }
 
     // -------------------------------------------------------------------------
-    // Record2 type implementation
+    // Record3 type implementation
     // -------------------------------------------------------------------------
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public Row2<Integer, String> fieldsRow() {
-        return (Row2) super.fieldsRow();
+    public Row3<Integer, String, String> fieldsRow() {
+        return (Row3) super.fieldsRow();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public Row2<Integer, String> valuesRow() {
-        return (Row2) super.valuesRow();
+    public Row3<Integer, String, String> valuesRow() {
+        return (Row3) super.valuesRow();
     }
 
     /**
@@ -95,7 +109,7 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      */
     @Override
     public Field<Integer> field1() {
-        return TokenTypes.TOKEN_TYPES.ID;
+        return JsonWebKey.JSON_WEB_KEY.ID;
     }
 
     /**
@@ -103,7 +117,15 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      */
     @Override
     public Field<String> field2() {
-        return TokenTypes.TOKEN_TYPES.TOKEN_TYPE;
+        return JsonWebKey.JSON_WEB_KEY.KEYID;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Field<String> field3() {
+        return JsonWebKey.JSON_WEB_KEY.JSON;
     }
 
     /**
@@ -119,7 +141,15 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      */
     @Override
     public String component2() {
-        return getTokenType();
+        return getKeyid();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String component3() {
+        return getJson();
     }
 
     /**
@@ -135,14 +165,22 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      */
     @Override
     public String value2() {
-        return getTokenType();
+        return getKeyid();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public TokenTypesRecord value1(Integer value) {
+    public String value3() {
+        return getJson();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWebKeyRecord value1(Integer value) {
         setId(value);
         return this;
     }
@@ -151,8 +189,8 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      * {@inheritDoc}
      */
     @Override
-    public TokenTypesRecord value2(String value) {
-        setTokenType(value);
+    public JsonWebKeyRecord value2(String value) {
+        setKeyid(value);
         return this;
     }
 
@@ -160,9 +198,19 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
      * {@inheritDoc}
      */
     @Override
-    public TokenTypesRecord values(Integer value1, String value2) {
+    public JsonWebKeyRecord value3(String value) {
+        setJson(value);
+        return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWebKeyRecord values(Integer value1, String value2, String value3) {
         value1(value1);
         value2(value2);
+        value3(value3);
         return this;
     }
 
@@ -171,19 +219,20 @@ public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> impl
     // -------------------------------------------------------------------------
 
     /**
-     * Create a detached TokenTypesRecord
+     * Create a detached JsonWebKeyRecord
      */
-    public TokenTypesRecord() {
-        super(TokenTypes.TOKEN_TYPES);
+    public JsonWebKeyRecord() {
+        super(JsonWebKey.JSON_WEB_KEY);
     }
 
     /**
-     * Create a detached, initialised TokenTypesRecord
+     * Create a detached, initialised JsonWebKeyRecord
      */
-    public TokenTypesRecord(Integer id, String tokenType) {
-        super(TokenTypes.TOKEN_TYPES);
+    public JsonWebKeyRecord(Integer id, String keyid, String json) {
+        super(JsonWebKey.JSON_WEB_KEY);
 
         set(0, id);
-        set(1, tokenType);
+        set(1, keyid);
+        set(2, json);
     }
 }
