+++ stroom-persistence/src/main/java/stroom/auth/db/tables/records/TokenTypesRecord.java
@@ -1,18 +1,18 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables.records;
+package stroom.auth.db.tables.records;
 
 
 import javax.annotation.Generated;
 
 import org.jooq.Field;
 import org.jooq.Record1;
-import org.jooq.Record3;
-import org.jooq.Row3;
+import org.jooq.Record2;
+import org.jooq.Row2;
 import org.jooq.impl.UpdatableRecordImpl;
 
-import stroom.db.auth.tables.JsonWebKey;
+import stroom.auth.db.tables.TokenTypes;
 
 
 /**
@@ -26,50 +26,36 @@ import stroom.db.auth.tables.JsonWebKey;
     comments = "This class is generated by jOOQ"
 )
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> implements Record3<Integer, String, String> {
+public class TokenTypesRecord extends UpdatableRecordImpl<TokenTypesRecord> implements Record2<Integer, String> {
 
-    private static final long serialVersionUID = 46155169;
+    private static final long serialVersionUID = 402226208;
 
     /**
-     * Setter for <code>auth.json_web_key.id</code>.
+     * Setter for <code>auth.token_types.id</code>.
      */
     public void setId(Integer value) {
         set(0, value);
     }
 
     /**
-     * Getter for <code>auth.json_web_key.id</code>.
+     * Getter for <code>auth.token_types.id</code>.
      */
     public Integer getId() {
         return (Integer) get(0);
     }
 
     /**
-     * Setter for <code>auth.json_web_key.keyId</code>.
+     * Setter for <code>auth.token_types.token_type</code>.
      */
-    public void setKeyid(String value) {
+    public void setTokenType(String value) {
         set(1, value);
     }
 
     /**
-     * Getter for <code>auth.json_web_key.keyId</code>.
+     * Getter for <code>auth.token_types.token_type</code>.
      */
-    public String getKeyid() {
+    public String getTokenType() {
         return (String) get(1);
-    }
-
-    /**
-     * Setter for <code>auth.json_web_key.json</code>.
-     */
-    public void setJson(String value) {
-        set(2, value);
-    }
-
-    /**
-     * Getter for <code>auth.json_web_key.json</code>.
-     */
-    public String getJson() {
-        return (String) get(2);
     }
 
     // -------------------------------------------------------------------------
@@ -85,23 +71,23 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
     }
 
     // -------------------------------------------------------------------------
-    // Record3 type implementation
+    // Record2 type implementation
     // -------------------------------------------------------------------------
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public Row3<Integer, String, String> fieldsRow() {
-        return (Row3) super.fieldsRow();
+    public Row2<Integer, String> fieldsRow() {
+        return (Row2) super.fieldsRow();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public Row3<Integer, String, String> valuesRow() {
-        return (Row3) super.valuesRow();
+    public Row2<Integer, String> valuesRow() {
+        return (Row2) super.valuesRow();
     }
 
     /**
@@ -109,7 +95,7 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      */
     @Override
     public Field<Integer> field1() {
-        return JsonWebKey.JSON_WEB_KEY.ID;
+        return TokenTypes.TOKEN_TYPES.ID;
     }
 
     /**
@@ -117,15 +103,7 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      */
     @Override
     public Field<String> field2() {
-        return JsonWebKey.JSON_WEB_KEY.KEYID;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Field<String> field3() {
-        return JsonWebKey.JSON_WEB_KEY.JSON;
+        return TokenTypes.TOKEN_TYPES.TOKEN_TYPE;
     }
 
     /**
@@ -141,15 +119,7 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      */
     @Override
     public String component2() {
-        return getKeyid();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String component3() {
-        return getJson();
+        return getTokenType();
     }
 
     /**
@@ -165,22 +135,14 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      */
     @Override
     public String value2() {
-        return getKeyid();
+        return getTokenType();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public String value3() {
-        return getJson();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public JsonWebKeyRecord value1(Integer value) {
+    public TokenTypesRecord value1(Integer value) {
         setId(value);
         return this;
     }
@@ -189,8 +151,8 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      * {@inheritDoc}
      */
     @Override
-    public JsonWebKeyRecord value2(String value) {
-        setKeyid(value);
+    public TokenTypesRecord value2(String value) {
+        setTokenType(value);
         return this;
     }
 
@@ -198,19 +160,9 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
      * {@inheritDoc}
      */
     @Override
-    public JsonWebKeyRecord value3(String value) {
-        setJson(value);
-        return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public JsonWebKeyRecord values(Integer value1, String value2, String value3) {
+    public TokenTypesRecord values(Integer value1, String value2) {
         value1(value1);
         value2(value2);
-        value3(value3);
         return this;
     }
 
@@ -219,20 +171,19 @@ public class JsonWebKeyRecord extends UpdatableRecordImpl<JsonWebKeyRecord> impl
     // -------------------------------------------------------------------------
 
     /**
-     * Create a detached JsonWebKeyRecord
+     * Create a detached TokenTypesRecord
      */
-    public JsonWebKeyRecord() {
-        super(JsonWebKey.JSON_WEB_KEY);
+    public TokenTypesRecord() {
+        super(TokenTypes.TOKEN_TYPES);
     }
 
     /**
-     * Create a detached, initialised JsonWebKeyRecord
+     * Create a detached, initialised TokenTypesRecord
      */
-    public JsonWebKeyRecord(Integer id, String keyid, String json) {
-        super(JsonWebKey.JSON_WEB_KEY);
+    public TokenTypesRecord(Integer id, String tokenType) {
+        super(TokenTypes.TOKEN_TYPES);
 
         set(0, id);
-        set(1, keyid);
-        set(2, json);
+        set(1, tokenType);
     }
 }
