+++ stroom-persistence/src/main/java/stroom/auth/db/tables/records/TokensRecord.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables.records;
+package stroom.auth.db.tables.records;
 
 
 import java.sql.Timestamp;
@@ -14,7 +14,7 @@ import org.jooq.Record11;
 import org.jooq.Row11;
 import org.jooq.impl.UpdatableRecordImpl;
 
-import stroom.db.auth.tables.Tokens;
+import stroom.auth.db.tables.Tokens;
 
 
 /**
@@ -30,7 +30,7 @@ import stroom.db.auth.tables.Tokens;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class TokensRecord extends UpdatableRecordImpl<TokensRecord> implements Record11<Integer, Integer, Integer, String, Timestamp, String, Timestamp, Integer, Boolean, Timestamp, Integer> {
 
-    private static final long serialVersionUID = 1299445033;
+    private static final long serialVersionUID = -1325069823;
 
     /**
      * Setter for <code>auth.tokens.id</code>.
