+++ stroom-persistence/src/main/java/stroom/auth/db/tables/records/UsersRecord.java
@@ -1,7 +1,7 @@
 /*
  * This file is generated by jOOQ.
  */
-package stroom.db.auth.tables.records;
+package stroom.auth.db.tables.records;
 
 
 import java.sql.Timestamp;
@@ -14,7 +14,7 @@ import org.jooq.Record17;
 import org.jooq.Row17;
 import org.jooq.impl.UpdatableRecordImpl;
 
-import stroom.db.auth.tables.Users;
+import stroom.auth.db.tables.Users;
 
 
 /**
@@ -30,7 +30,7 @@ import stroom.db.auth.tables.Users;
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class UsersRecord extends UpdatableRecordImpl<UsersRecord> implements Record17<Integer, String, String, Timestamp, String, String, String, String, Integer, Integer, Timestamp, Timestamp, String, Timestamp, String, Boolean, Timestamp> {
 
-    private static final long serialVersionUID = 1974101103;
+    private static final long serialVersionUID = 1836991455;
 
     /**
      * Setter for <code>auth.users.id</code>.
